Refactor ms-audit service
- Removed unused environment variables from deployment.yaml.
- Updated deployment probes to use named ports.
- Refactored main.go to remove unused imports and initialize health server.
- Updated go.mod to reflect version changes and remove unused dependencies.
- Cleaned up go.sum by removing unnecessary entries.
- Refactored settings.go to remove versioning and vault secret handling.
- Deleted obsolete RabbitMQ definitions and handler files.
- Introduced new handler structure for RabbitMQ messages.
- Updated listener logic to use new handler structure.
- Removed unused HTTP server initialization code.
- Updated connection initialization to use new RabbitMQ connection method.
