+++ charts/ror-ms-audit/templates/deployment.yaml
@@ -36,8 +36,6 @@ spec:
           env:
             - name: LOG_LEVEL
               value: {{ default "INFO" .Values.nhn.logLevel | quote }}
-            - name: ENVIRONMENT
-              value: {{ default "Production" .Values.nhn.environment | quote }}
             - name: RABBITMQ_HOST
               value: {{ default "rabbitmq-ror.nhn-ror.svc" .Values.nhn.rabbitmqHost | quote }}
             - name: RABBITMQ_PORT
@@ -52,12 +50,6 @@ spec:
               value: {{ default "ror-ms-audit" .Values.nhn.role | quote }}
             - name: OPENTELEMETRY_COLLECTOR_ENDPOINT
               value: "ror-opentelemetry-collector.nhn-ror.svc:4317"
-            - name: MONGODB_PORT
-              value: {{ default "27017" .Values.nhn.mongodbPort | quote }}
-            - name: MONGODB_DATABASE
-              value: {{ default "nhn-ror" .Values.nhn.mongodbDatabase | quote }}
-            - name: MONGODB_HOST
-              value: {{ default "ror-mongodb.nhn-ror.svc" .Values.nhn.mongodbHost | quote}}
           ports:
             - name: http
               containerPort: 8080
@@ -65,11 +57,11 @@ spec:
           livenessProbe:
             httpGet:
               path: /health
-              port: 8080
+              port: http
           readinessProbe:
             httpGet:
               path: /health
-              port: 8080
+              port: http
           resources:
             {{- toYaml .Values.resources | nindent 12 }}
       {{- with .Values.nodeSelector }}
